Details: Modified pause_expansion() behaviour to cope with infinite blocking of main thread Added get_size set_expansion_rate() added for simpler access to modify the rate (not much change to logic, just abstracted it) Added initiliaser based constructor (this also resize already provided logic) MOST NOTICIABLE IS - Completed Swastic_Find

+++ matrix_base.hpp
@@ -14,7 +14,7 @@ protected:	// it shouldn't be instantiable
 
 	int getDimensionsOfMatrix() { return this->_dimension; }
 
-	static constexpr uint8_t init_expansion_speed{ 1 }; // @future - allow negative growths too
+	static constexpr float init_expansion_speed{ 1 }; // @future - allow negative growths too
 	virtual void auto_expansion() = 0;
 	virtual void pause_auto_expansion() = 0;
 	virtual void resume_auto_expansion() = 0;
